requirements.txt
# Updated requirements for FastCP
# Core Django and web framework
django>=5.1.1
djangorestframework==3.16.1
asgiref==3.9.1
uvicorn==0.35.0
gunicorn==23.0.0

# Database
mysqlclient==2.2.7
dj-database-url==2.2.0

# Security and certificates
acme==2.11.0  # Keeping current version due to major version jump (2.x -> 5.x)
certifi==2025.8.3
cryptography>=44.0.1
pyOpenSSL>=24.3.0
josepy==1.14.0

# HTTP requests and networking
requests>=2.32.4
requests-toolbelt==1.0.0
urllib3>=2.5.0
h11>=0.16.0

# Character encoding and validation
chardet==5.2.0
charset-normalizer==3.4.3
idna==3.10
validators==0.35.0

# Async and process management
gevent>=25.4.2
greenlet>=3.2.0
psutil==6.0.0  # Keeping current version due to major version jump (6.x -> 7.x)

# Utilities and helpers
django-common-helpers==0.9.2
click==8.2.1
decorator==5.2.1
six==1.17.0
whitenoise==6.9.0

# Date and time handling
python-dateutil==2.9.0.post0
pytz==2025.2
pyRFC3339==1.1

# Development and code formatting
autopep8==2.3.2
pycodestyle==2.14.0
toml==0.10.2

# Process management
pexpect==4.9.0
ptyprocess==0.7.0

# Cron functionality
python-crontab==3.3.0

# Configuration parsing
cffi==1.17.1
pycparser==2.22

# Event handling
zope.event==5.1.1
zope.interface==7.2

# SQL parsing
sqlparse==0.5.3

# Docker integration
docker==6.0.0
docker-pycreds==0.4.0
